backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pdfplumber==0.11.4
google-generativeai==0.8.3
openai>=1.0.0
chromadb==0.5.23
numpy==2.2.1
python-dotenv==1.0.1
python-multipart==0.0.20
firebase-admin==6.6.0
pydantic==2.10.4
httpx>=0.25.0
